Split the store's till and drawings from its storefront
The cart readout, the board and sheet drawings, and the trip's till hook
are about shopping, not about the menu they happen to be laid out in
today — the lumberyard already reached across for two of them, and a
store you walk around wants the same three for its shelf tags.

They move to shopping/. What stays in store-page/ is the storefront:
the aisle sections, the product tiles, and the overlay that arranges
them, which is the part that becomes the Orange Box website (#200).

Deliberately narrow: only the pieces that are certainly shared however
Phase 1 lands. BoardSelector stays put until the rack card actually
needs it.
