+++ src/components/shopping/BoardFaceSvg.tsx
