+++ src/components/store-page/StoreSuppliesSection.tsx
@@ -12,7 +12,7 @@ import { ClampIcon, ConsumableIcon } from "../ItemIcon";
 import { useApplyGameAction, useGameState } from "../useGameState";
 import { AisleSection } from "./AisleSection";
 import { ProductTile } from "./ProductTile";
-import { useCartCount } from "./useStoreTrip";
+import { useCartCount } from "../shopping/useStoreTrip";
 
 export const StoreSuppliesSection: React.FC<{ className?: string }> = ({
   className,
