+++ src/components/store-page/StoreTripOverlay.tsx
@@ -3,8 +3,8 @@ import { TripHeader } from "../trip/TripHeader";
 import { TripOverlay } from "../trip/TripOverlay";
 import { useGameState } from "../useGameState";
 import { BoardSelector } from "./BoardSelector";
-import { StoreCartReadout, StoreCheckoutButton } from "./StoreCart";
-import { useStoreTrip } from "./useStoreTrip";
+import { StoreCartReadout, StoreCheckoutButton } from "../shopping/StoreCart";
+import { useStoreTrip } from "../shopping/useStoreTrip";
 import { StoreMachinesSection } from "./StoreMachinesSection";
 import { StoreSheetGoodsSection } from "./StoreSheetGoodsSection";
 import { StoreSuppliesSection } from "./StoreSuppliesSection";
